Wrap op-heavy bench setup in try_setup for tolerant skipping
bench_synced only catches panics inside the timed closure. Setup ops that run before
it (make_qtensor calls quantize_dynamic; fft_ops' irfft bench calls rfft to produce
the input; make_int_tensor hits tch's int_from_data which only supports i64) panicked
and took down the whole bench binary.

common::try_setup runs a setup closure under catch_unwind, records the file:line and
panic message, and returns None on failure. Updated:
  - make_qtensor / make_qmatrix return Option (quantization on tch, others)
  - make_int_tensor / make_int_tensor_2d / make_indices* return Option
  - fft_ops' irfft helper wraps its rfft input generation

Also switched remaining Vec<i64> bench data to Vec<i32> to match IntElem. i64 range
in Tensor::arange stays i64 (required by API).

+++ crates/burn-backend-tests/benches/int_ops.rs
@@ -26,7 +26,7 @@ fn main() {
 fn make_int_tensor<B: Backend>(shape: &[usize]) -> Tensor<B, 2, Int> {
     let size: usize = shape.iter().product();
     // Use values that fit in i8 (-128 to 127) so all casts work
-    let data: Vec<i64> = (0..size).map(|i| (i % 200) as i64 - 100).collect();
+    let data: Vec<i32> = (0..size).map(|i| (i % 200) as i32 - 100).collect();
     Tensor::from_data(TensorData::new(data, shape.to_vec()), &Default::default())
 }
